fix: [hior-1] tests
