+++ feature/settings/ui/src/main/java/com/grappim/hateitorrateit/feature/settings/ui/screen/backup/SettingsBackupScreen.kt
@@ -82,7 +82,7 @@ fun SettingsBackupRoute(viewModel: SettingsBackupViewModel = hiltViewModel()) {
 
     state.lastImportResult?.let { importResult ->
         if (state.shouldShowImportResultDialog) {
-            ImportResultDialog(
+            ImportResultDialogWidget(
                 importResult = importResult,
                 onDismiss = state.onImportResultDialogDismissed
             )
@@ -337,208 +337,6 @@ private fun ImportModeDialog(onImportModeSelect: (ImportMode) -> Unit, onDismiss
         dismissButton = {
             TextButton(onClick = onDismiss) {
                 Text(stringResource(RString.cancel))
-            }
-        }
-    )
-}
-
-@Composable
-private fun ImportResultDialog(importResult: ImportResult, onDismiss: () -> Unit) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Icon(
-                    imageVector = when (importResult) {
-                        is ImportResult.Success -> Icons.Filled.CheckCircle
-                        is ImportResult.PartialSuccess -> Icons.Filled.Warning
-                        is ImportResult.Failure -> Icons.Filled.Error
-                    },
-                    contentDescription = null,
-                    tint = when (importResult) {
-                        is ImportResult.Success -> MaterialTheme.colorScheme.primary
-                        is ImportResult.PartialSuccess -> MaterialTheme.colorScheme.secondary
-                        is ImportResult.Failure -> MaterialTheme.colorScheme.error
-                    }
-                )
-                Text(
-                    text = when (importResult) {
-                        is ImportResult.Success -> stringResource(
-                            RString.import_completed_successfully
-                        )
-                        is ImportResult.PartialSuccess -> stringResource(
-                            RString.import_completed_with_warnings
-                        )
-                        is ImportResult.Failure -> stringResource(RString.import_failed)
-                    }
-                )
-            }
-        },
-        text = {
-            LazyColumn(
-                modifier = Modifier.height(400.dp)
-            ) {
-                when (importResult) {
-                    is ImportResult.Success -> {
-                        item {
-                            Column {
-                                Text(
-                                    text = stringResource(RString.successfully_imported),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.primary
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Text(
-                                    "• ${stringResource(
-                                        RString.products_count,
-                                        importResult.importedProducts
-                                    )}"
-                                )
-                                Text(
-                                    "• ${stringResource(
-                                        RString.images_count,
-                                        importResult.importedImages
-                                    )}"
-                                )
-                                if (importResult.importedSettings) {
-                                    Text("• ${stringResource(RString.app_settings)}")
-                                }
-                            }
-                        }
-                    }
-
-                    is ImportResult.PartialSuccess -> {
-                        item {
-                            Column {
-                                Text(
-                                    text = stringResource(RString.successfully_imported),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.primary
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Text(
-                                    "• ${stringResource(
-                                        RString.products_count,
-                                        importResult.importedProducts
-                                    )}"
-                                )
-                                Text(
-                                    "• ${stringResource(
-                                        RString.images_count,
-                                        importResult.importedImages
-                                    )}"
-                                )
-                                if (importResult.importedSettings) {
-                                    Text("• ${stringResource(RString.app_settings)}")
-                                }
-
-                                Spacer(modifier = Modifier.height(16.dp))
-                            }
-                        }
-
-                        if (importResult.skippedProducts.isNotEmpty()) {
-                            item {
-                                Text(
-                                    text = stringResource(RString.skipped_products),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.onSurface
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                            }
-                            items(importResult.skippedProducts) { productName ->
-                                Text(
-                                    text = "• $productName",
-                                    style = MaterialTheme.typography.bodySmall,
-                                    modifier = Modifier.padding(bottom = 4.dp)
-                                )
-                            }
-                            item { Spacer(modifier = Modifier.height(8.dp)) }
-                        }
-
-                        if (importResult.failedImages.isNotEmpty()) {
-                            item {
-                                Text(
-                                    text = stringResource(RString.failed_images),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.error
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                            }
-                            items(importResult.failedImages) { imageName ->
-                                Text(
-                                    text = "• $imageName",
-                                    style = MaterialTheme.typography.bodySmall,
-                                    modifier = Modifier.padding(bottom = 4.dp),
-                                    color = MaterialTheme.colorScheme.error
-                                )
-                            }
-                            item { Spacer(modifier = Modifier.height(8.dp)) }
-                        }
-
-                        if (importResult.warnings.isNotEmpty()) {
-                            item {
-                                Text(
-                                    text = stringResource(RString.warnings),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.onSurface
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                            }
-                            items(importResult.warnings) { warning ->
-                                Card(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .padding(bottom = 4.dp),
-                                    shape = RoundedCornerShape(8.dp)
-                                ) {
-                                    Text(
-                                        text = warning,
-                                        style = MaterialTheme.typography.bodySmall,
-                                        modifier = Modifier.padding(8.dp)
-                                    )
-                                }
-                            }
-                        }
-                    }
-
-                    is ImportResult.Failure -> {
-                        item {
-                            Column {
-                                Text(
-                                    text = stringResource(RString.import_failed_message),
-                                    style = MaterialTheme.typography.titleSmall,
-                                    color = MaterialTheme.colorScheme.error
-                                )
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Card(
-                                    modifier = Modifier.fillMaxWidth(),
-                                    shape = RoundedCornerShape(8.dp)
-                                ) {
-                                    Text(
-                                        text = importResult.message,
-                                        style = MaterialTheme.typography.bodySmall,
-                                        modifier = Modifier.padding(12.dp),
-                                        color = MaterialTheme.colorScheme.error
-                                    )
-                                }
-                                Spacer(modifier = Modifier.height(8.dp))
-                                Text(
-                                    text = stringResource(RString.error_type, importResult.error),
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringResource(RString.ok))
             }
         }
     )
